insecure debug endpoints only in testing (#91) - /debug/tasks, /debug/index and /debug/sign are only compiled in test mode
